backend/requirements.txt
fastapi
uvicorn
langchain
langchain-groq
python-dotenv
wikipedia
langchain-community
langchain_tavily